Ordenamiento de código
todas las funciones se fueron pa redchaucha.py y en el main se quedaron las llamadas desde el bot de telegram
