+++ __main__.py
@@ -1,94 +1,20 @@
 from telegram.ext import Updater, CommandHandler, MessageHandler, Filters
-from requests import get, post
-from bitcoin import *
-from config import *
+from config import token
+from redchaucha import *
 import logging
 
 logging.basicConfig(format='%(asctime)s - %(levelname)s - %(message)s', level=logging.INFO)
 logger = logging.getLogger(__name__)
 
-def getaddress(user_id):
-	privkey = sha256(str(user_id) + str(salt))
-	addr = privtoaddr(privkey, magic)
-	return [addr, privkey]
-
-def getbalance(addr):
-	# Captura de balance por tx sin gastar
-	unspent = get('https://explorer.cha.terahash.cl/api/addr/' + addr + '/utxo').json()
-		
-	confirmed = unconfirmed = 0
-
-	inputs = []
-	for i in unspent:
-		if i['confirmations'] >= 6:
-			confirmed += i['amount']
-			inputs_tx = {'output' : i['txid'] + ':' + str(i['vout']), 'value' : i['satoshis'], 'address' : i['address']}
-			inputs.append(inputs_tx)
-		else:
-			unconfirmed += i['amount']
-
-	return [confirmed, inputs, unconfirmed]
-
-
 def send(bot, update, args):
 	try:
 		user = update.message.from_user
-
+		
 		amount = float(args[0])
 		receptor = args[1]
 
-		info = getaddress(user.id)
-		addr = info[0]
-		privkey = info[1]
-		
-		info = getbalance(addr)
-		confirmed_balance = info[0]
-		inputs = info[1]
+		msg = sendTx(user.id, amount, receptor)
 
-		if not len(receptor) == 34 and receptor[0] == 'c':
-			msg = "Dirección inválida"
-
-		elif not confirmed_balance - fee >= amount:
-			msg = "Balance insuficiente"
-
-		elif not amount > 0:
-			msg = "Monto inválido"
-
-		else:
-			# Transformar valores a Chatoshis
-			amount = int(amount*satoshi)
-			used_fee = int(fee*satoshi)
-
-			# Utilizar solo las unspent que se necesiten
-			used_balance = 0
-			used_inputs = []
-
-			for i in inputs:
-				used_balance += i['value']
-				used_inputs.append(i)
-				if used_balance >= amount:
-					break 
-
-			# Creación de salida
-			outputs = [{'address' : receptor, 'value' : amount}]
-
-			# Agregar transacción de vuelto si es necesario
-			if not int(used_balance - amount) == 0:
-				outputs.append({'address' : addr, 'value' : int(used_balance - (amount + used_fee))}) 
-
-			# Transacción
-			tx = mktx(used_inputs, outputs)
-
-			# Firma
-			for i in range(len(used_inputs)):
-				tx = sign(tx, i, privkey)
-
-			broadcasting = post('https://explorer.cha.terahash.cl/api/tx/send', data={'rawtx' : tx})
-
-			try:
-				msg = "insight.chaucha.cl/tx/%s" % broadcasting.json()['txid']
-			except:
-				msg = "ERROR: %s" % broadcasting.text()
 	except:
 		msg = "Error de formato >:C\n\n"
 		msg += "Modo de uso: /send monto address"
@@ -137,8 +63,9 @@ def main():
 	# log all errors
 	dp.add_error_handler(error)
 
+	quirquincho = getaddress('Quirquincho')[0]
 	# Inicio de bot
-	logger.info("Quirquincho V 2.0")
+	logger.info("Quirquincho V 2.0 - %s" % quirquincho)
 	updater.start_polling()
 
 	updater.idle()
